# garden-app/api/openapi.yaml
openapi: 3.0.3
info:
  description: |-
    This is an OpenAPI Specification for https://github.com/calvinmclean/automated-garden/tree/main/garden-app

    It allows managing and interacting with Gardens and Plants for this application.

  title: Garden App - OpenAPI
  version: 1.0.0
...
paths:
...
  /gardens/{gardenID}/zones:
    post:
      tags:
        - zones
      summary: Add a Zone
      description: Adds a new Zone to this Garden.
      operationId: addZone
      parameters:
        - $ref: "#/components/parameters/GardenID"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZoneResponse"
        "400":
          description: Bad Request
      requestBody:
        description: Add a Zone
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateZoneRequest"
    get:
      tags:
        - zones
      summary: Get all Zones
      description: Query for a list of all Zones. Optionally include end-dated Zones.
      operationId: getAllZones
      parameters:
        - $ref: "#/components/parameters/GardenID"
        - name: end_dated
          in: query
          description: whether or not to include end-dated Zones in the response
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AllZonesResponse"
        "400":
          description: Bad Request
  /gardens/{gardenID}/zones/{zoneID}:
    get:
      tags:
        - zones
      summary: Get a Zone
      description: Get details of a Zone.
      operationId: getZone
      parameters:
        - $ref: "#/components/parameters/GardenID"
        - $ref: "#/components/parameters/ZoneID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZoneResponse"
        "400":
          description: Bad Request
    patch:
      tags:
        - zones
      summary: Update/Edit a Zone
      description: Update/Edit a Zone. Only certain fields of a Zone are editable.
      operationId: updateZone
      parameters:
        - $ref: "#/components/parameters/GardenID"
        - $ref: "#/components/parameters/ZoneID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZoneResponse"
        "400":
          description: Bad Request
      requestBody:
        description: Update/Edit a new Zone
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateZoneRequest"
    delete:
      tags:
        - zones
      summary: End-date a Zone
      description: End-date a Zone. This allows deleting without actually losing the resource data. Once a Zone is end-dated, certain resources will not be available such as health and action.
      operationId: endDateZone
      parameters:
        - $ref: "#/components/parameters/GardenID"
        - $ref: "#/components/parameters/ZoneID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZoneResponse"
        "400":
          description: Bad Request
  /gardens/{gardenID}/zones/{zoneID}/action:
    post:
      tags:
        - zones
      summary: Execute action on a Zone
      description: This endpoint allows executing certain actions on a Zone. This will actually interact with the `garden-controller` rather than just manipulating resources.
      operationId: zoneAction
      parameters:
        - $ref: "#/components/parameters/GardenID"
        - $ref: "#/components/parameters/ZoneID"
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request
      requestBody:
        description: Execute action on a Zone
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZoneAction"
  /gardens/{gardenID}/zones/{zoneID}/history:
    post:
      tags:
        - zones
      summary: Get Zone's watering history
      description: This endpoint retrieves information about past watering events for this Zone
      operationId: zoneHistory
      parameters:
        - $ref: "#/components/parameters/GardenID"
        - $ref: "#/components/parameters/ZoneID"
        - name: range
          in: query
          description: duration describing the amount of time in the past to show events from (default=72h)
          required: false
          schema:
            type: string
            example: 72h
        - name: limit
          in: query
          description: maximum number of events to include in response (default=0/no limit)
          required: false
          schema:
            type: integer
            example: 5
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaterHistoryResponse"
        "400":
          description: Bad Request
components:
  parameters:
    GardenID:
      name: gardenID
      in: path
      description: ID of Garden resource for this request
      required: true
      schema:
        $ref: "#/components/schemas/xid"
...
    ZoneID:
      name: zoneID
      in: path
      description: ID of Zone resource for this request
      required: true
      schema:
        $ref: "#/components/schemas/xid"
  schemas:
    xid:
      type: string
      description: 20 character unique identifier used throughout this API
      externalDocs:
        url: https://github.com/rs/xid
      pattern: "[0-9a-v]{20}"
      example: 9m4e2mr0ui3e8a215n4g
    link:
      type: object
      description: used for JSON Hypermedia to link to additional resources
      properties:
        ref:
          type: string
          description: one-word explanation of what this link is for
        href:
          type: string
          description: relative path for the specified resource
...
    WaterSchedule:
      type: object
      description: |
        Allows control over how/when the Zone is watered. The `interval` is used to control the amount of time the pump runs to water the Garden.
        Specifying a `minimum_moisture` will instruct the server to consider moisture sensor data when executing the WaterAction.
        To avoid falsely triggering watering and avoid the complexity if constantly checking moisture data, the moisture is only checked when executing a WaterAction.
      properties:
        duration:
          type: string
          description: amount of time, in Duration format, that Plant should be watered
          example: 15000ms
        interval:
          type: string
          format: duration
          description: amount of time, as a Duration string, to wait between scheduled watering
          example: 72h
        minimum_moisture:
          type: integer
          minimum: 0
          maximum: 100
          description: this is a percentage representing the threshold that the Plant's moisture must be below to enable a WaterAction
          example: 50
        start_time:
          type: string
          format: date-time
          description: time that the watering interval should be started at
      required:
        - duration
        - interval
        - start_time
    Zone:
      type: object
      description: |
        A Zone corresponds to a "waterable resource" and enabled direct control of `garden-controller` watering devices.
        It holds a WaterSchedule and can be referenced by Plants to track their watering and location. This allows watering
        multiple Plants from a single control and is useful for larger Garden setups, such as an outdoor Garden with beds.
      properties:
        name:
          type: string
          description: this is the name of the Zone
          example: Raised Bed 1
        details:
          type: object
          description: this holds various additional information about a Zone that might be useful to remember, but is not related to core functionality of this application.
          properties:
            description:
              type: string
              description: a general description of a Zone
              example: raised bed containing assorted vegetables
            notes:
              type: string
              description: general notes about the Zone
              example: near the west wall of the yard, receiving morning light and afternoon shade
        position:
          type: integer
          description: this corresponds directly to the configuration of the `garden-controller` and is critical for having control of the Zone
          example: 0
          minimum: 0
        water_schedule:
          $ref: "#/components/schemas/WaterSchedule"
    UpdateZoneRequest:
      type: object
      description: This allows updating/editing a Zone resource
      allOf:
        - $ref: "#/components/schemas/Zone"
        - properties:
            created_at:
              type: string
              format: date-time
              description: the date-time when the Zone was originally created
    AllZonesResponse:
      type: object
      description: List of all Zones
      properties:
        plants:
          type: array
          items:
            $ref: "#/components/schemas/ZoneResponse"
    CreateZoneRequest:
      type: object
      description: This allows creating a Zone resource
      allOf:
        - $ref: "#/components/schemas/Zone"
      required:
        - name
        - position
        - water_schedule
    ZoneResponse:
      type: object
      description: This is the response object for Zones that contains extra information only available on Zones that are created
      allOf:
        - $ref: "#/components/schemas/Zone"
        - properties:
            id:
              $ref: "#/components/schemas/xid"
            created_at:
              type: string
              format: date-time
              description: the date-time when the Zone was originally created
            end_date:
              type: string
              format: date-time
              description: the date-time when the Zone was deleted/removed
            next_water_time:
              type: string
              description: time of the next scheduled watering
              format: date-time
            moisture:
              type: number
              format: float
              description: moisture percentage of a Zone with a soil moisture sensor
            links:
              type: array
              items:
                $ref: "#/components/schemas/link"
              example:
                - rel: self
                  href: /gardens/c22tmvucie6n6gdrpal0/plants/c3ucvu06n88pt1dom670
                - rel: garden
                  href: /gardens/c22tmvucie6n6gdrpal0
                - rel: action
                  href: /gardens/c22tmvucie6n6gdrpal0/plants/c3ucvu06n88pt1dom670/action
                - rel: history
                  href: /gardens/c22tmvucie6n6gdrpal0/plants/c3ucvu06n88pt1dom670/history
      required:
        - id
        - created_at
        - links
        - name
        - position
        - water_schedule
    WaterHistoryResponse:
      type: object
      description: response containing a list of past watering events and some basic aggegrate details about them
      properties:
        history:
          type: array
          items:
            $ref: "#/components/schemas/WaterHistory"
        count:
          type: integer
          description: number of watering events found in the range
          example: 1
        average:
          type: string
          description: average of `duration` for all events found. Formatted as a float in Go duration format
          example: 15s
        total:
          type: string
          description: total of `duration` for all events found. Formatted as a float in Go duration format
          example: 15s
    WaterHistory:
      type: object
      description: a
      properties:
        duration:
          type: string
          description: amount of time, in Duration format, that Zone should be watered
          example: 15000ms
        record_time:
          type: string
          format: date-time
          description: time that the watering event was recorded
    ZoneAction:
      type: object
      description: collects all the possible actions for a Zone into a single struct so these can easily be received as one request
      properties:
        water:
          $ref: "#/components/schemas/WaterAction"
      required:
        - water
    WaterAction:
      type: object
      description: waters a Zone for the specified amount of time
      properties:
        duration:
          type: number
          description: amount of time, in milliseconds, that Zone should be watered
          example: 150000
        ignore_moisture:
          type: boolean
          description: if Zone is configured with a `minimum_moisture` for watering, ignore it and force watering
      required:
        - duration